# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: RunBuoy API
  version: 1.2.0
  description: >-
    One-way synchronization API. Machines may write safe execution events;
    devices may read projections and manage only their own receiving plane.
servers:
  - url: https://api.example.runbuoy.dev
paths:
...
  /v1/plugin-connections:
    get:
      operationId: listPluginConnections
      summary: List active read-only connections confirmed by this device
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Device-owned connections; no token material
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PluginConnection' }
        '403': { description: Device identity and enrolled workspace required }
        '404': { description: Private preview disabled }
  /v1/plugin-connections/{connection_id}:
    delete:
      operationId: revokePluginConnection
      summary: Revoke this device's connection and all its OAuth tokens
      security: [{ bearerAuth: [] }]
      parameters:
        - name: connection_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Connection revoked }
        '404': { description: Connection is not owned by this device }
  /v1/plugin-connections/requests/{request_id}/inspect:
    post:
      operationId: inspectPluginConnection
      summary: Inspect a QR challenge without granting access
      security: [{ bearerAuth: [] }]
      parameters:
        - name: request_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PluginConnectionChallenge' }
      responses:
        '200':
          description: Server-verified client, workspace, read scopes and expiry for consent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginConnectionRequest' }
        '404': { description: Request or challenge not found }
        '410': { description: Five-minute request expired }
  /v1/plugin-connections/requests/{request_id}/decision:
    post:
      operationId: decidePluginConnection
      summary: Explicitly allow or deny a connection after phone consent
      security: [{ bearerAuth: [] }]
      parameters:
        - name: request_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, decision]
              additionalProperties: false
              properties:
                challenge: { type: string, minLength: 16, maxLength: 256 }
                decision: { enum: [allow, deny] }
      responses:
        '204': { description: Decision recorded once; return to the authorization browser }
        '403': { description: Enrolled device required }
        '404': { description: Request or challenge not found }
        '409': { description: Request already decided }
        '410': { description: Request expired }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    PluginConnectionChallenge:
      type: object
      required: [challenge]
      additionalProperties: false
      properties:
        challenge: { type: string, minLength: 16, maxLength: 256 }
    PluginConnection:
      type: object
      required: [id, client_name, scopes, created_at, expires_at]
      properties:
        id: { type: string }
        client_name: { type: string }
        scopes:
          type: array
          items: { enum: ['runs:read', 'machines:read', 'notifications:read'] }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
    PluginConnectionRequest:
      type: object
      required: [id, client_name, origin, workspace_id, scopes, status, expires_at]
      properties:
        id: { type: string }
        client_name: { type: string }
        origin: { type: string, format: uri }
        workspace_id: { type: string }
        scopes:
          type: array
          items: { enum: ['runs:read', 'machines:read', 'notifications:read'] }
        status: { enum: [pending, allowed, denied] }
        expires_at: { type: string, format: date-time }
...